backend_v2/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
bcrypt==4.0.1
python-multipart>=0.0.6
requests>=2.31.0
psycopg2-binary>=2.9.9

# Essential ML dependencies
numpy>=1.24.3

# RAG and similarity search dependencies
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.18
rank-bm25>=0.2.2
